content/canvas/test/test_2d.gradient.object.compare.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.gradient.object.compare</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
var _deferred = false;

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var g1 = ctx.createLinearGradient(0, 0, 100, 0);
var g2 = ctx.createLinearGradient(0, 0, 100, 0);
ok(g1 !== g2, "g1 !== g2");
ctx.fillStyle = g1;
ok(ctx.fillStyle === g1, "ctx.fillStyle === g1");

if (!_deferred) SimpleTest.finish();
});
</script>
